+++ Editor/Blend2DAttributeDrawer.cs
@@ -1,10 +1,6 @@
-﻿using System;
-using UnityEngine;
+﻿using UnityEngine;
 using UnityEditor;
-using System.Reflection;
 using UnityEditor.UIElements;
-using UnityEngine.Profiling;
-using UnityEngine.Rendering;
 using UnityEngine.UIElements;
 #if UNITY_MATHEMATICS
 using Unity.Mathematics;
@@ -12,7 +8,6 @@ using Unity.Mathematics;
 
 namespace Vertx.Attributes.Editor
 {
-#if UNITY_2022_1_OR_NEWER
 	public sealed class Blend2DBoxElement : VisualElement, IBindable
 	{
 		public const string BoxUssStyleName = Blend2DAttributeDrawer.UssStyleName + "__box";
@@ -171,16 +166,15 @@ namespace Vertx.Attributes.Editor
 			painter2D.Stroke();
 		}
 
-		public IBinding binding { get; set; }
-		public string bindingPath { get; set; }
+		public IBinding? binding { get; set; }
+		public string? bindingPath { get; set; }
 	}
-#endif
 	
 	[CustomPropertyDrawer(typeof(Blend2DAttribute))]
 	public sealed class Blend2DAttributeDrawer : PropertyDrawer
 	{
-		private const float blendBoxSize = 151f;
-		public const float CircleRadius = blendBoxSize * 0.04f;
+		private const float BlendBoxSize = 151f;
+		public const float CircleRadius = BlendBoxSize * 0.04f;
 
 		public static Color LowGrey
 		{
@@ -192,9 +186,8 @@ namespace Vertx.Attributes.Editor
 			}
 		}
 
-		public static Color CircleColor => new Color(1, 0.5f, 0);
-
-#if UNITY_2022_1_OR_NEWER
+		public static Color CircleColor => new(1, 0.5f, 0);
+		
 		public const string UssStyleName = "vertx-blend-2d";
 		public const string RightUssStyleName = UssStyleName + "__right";
 		public const string LabelUssStyleName = UssStyleName + "__label";
@@ -203,7 +196,7 @@ namespace Vertx.Attributes.Editor
 
 		public override VisualElement CreatePropertyGUI(SerializedProperty property)
 		{
-			Blend2DAttribute b2D = (Blend2DAttribute)attribute;
+			var b2D = (Blend2DAttribute)attribute;
 			var xLimit = new Vector2(b2D.Min.x, b2D.Max.x);
 			var yLimit = new Vector2(b2D.Min.y, b2D.Max.y);
 			SerializedProperty x = property.FindPropertyRelative("x");
@@ -230,6 +223,8 @@ namespace Vertx.Attributes.Editor
 			AddProperty(b2D.XLabel, x, xLimit);
 			AddProperty(b2D.YLabel, y, yLimit);
 
+			return root;
+
 			void AddProperty(string text, SerializedProperty p, Vector2 limits)
 			{
 				// This setup lets us have a non-delayed dragger, and a delayed float field.
@@ -250,6 +245,7 @@ namespace Vertx.Attributes.Editor
 
 				fieldDragger.RegisterCallback<ChangeEvent<float>, Vector2>(Clamp, limits);
 				field.RegisterCallback<ChangeEvent<float>, Vector2>(Clamp, limits);
+				return;
 
 				void Clamp(ChangeEvent<float> evt, Vector2 args)
 				{
@@ -260,190 +256,12 @@ namespace Vertx.Attributes.Editor
 						floatField.SetValueWithoutNotify(args.y);
 				}
 			}
-
-			return root;
-		}
-#endif
-
-		public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
-		{
-			Blend2DAttribute b2D = (Blend2DAttribute)attribute;
-			using (new EditorGUI.PropertyScope(position, GUIContent.none, property))
-				Do2DBlend(position, property, b2D);
 		}
 
-		private int hotControl = -1;
-		private readonly int hash = "Blend2D".GetHashCode();
-
-		private void Do2DBlend(Rect r, SerializedProperty property, Blend2DAttribute b2D)
-		{
-			Profiler.BeginSample(nameof(Do2DBlend));
-			GUI.BeginGroup(r);
-			try
-			{
-				Rect blendRect = new Rect(Vector2.zero, Vector2.one * blendBoxSize);
-
-				Event e = Event.current;
-				if (e.isMouse && e.button == 0 && e.rawType == EventType.MouseDown)
-				{
-					if (blendRect.Contains(e.mousePosition))
-					{
-						hotControl = GUIUtility.GetControlID(hash, FocusType.Passive, blendRect);
-						GUIUtility.hotControl = hotControl;
-						e.Use();
-					}
-				}
-
-				if (e.isMouse && e.rawType == EventType.MouseUp)
-				{
-					switch (e.button)
-					{
-						case 0:
-							hotControl = -1;
-							GUIUtility.hotControl = 0;
-							e.Use();
-							break;
-						case 1:
-							if (!blendRect.Contains(e.mousePosition))
-								break;
-							SerializedProperty x = property.FindPropertyRelative("x");
-							SerializedProperty y = property.FindPropertyRelative("y");
-							ShowContextMenu(
-								x, y,
-								new Vector2(b2D.Min.x, b2D.Max.x),
-								new Vector2(b2D.Min.y, b2D.Max.y)
-							);
-							e.Use();
-							break;
-					}
-				}
-
-				if (GUIUtility.hotControl == hotControl)
-				{
-					if (e.isMouse)
-					{
-						Vector2 value = Lerp(
-							b2D.Min,
-							b2D.Max,
-							new Vector2(e.mousePosition.x / blendBoxSize, (blendBoxSize - e.mousePosition.y) / blendBoxSize));
-#if UNITY_MATHEMATICS
-						if (IsFloat2())
-						{
-							SerializedProperty x = property.FindPropertyRelative("x");
-							SerializedProperty y = property.FindPropertyRelative("y");
-							x.floatValue = value.x;
-							y.floatValue = value.y;
-						}
-						else
-#endif
-						{
-							property.vector2Value = value;
-						}
-
-						e.Use();
-					}
-				}
-
-				GUI.Box(blendRect, "", EditorStyles.helpBox);
-
-				if (e.type == EventType.Repaint)
-				{
-					GL.Begin(Application.platform == RuntimePlatform.WindowsEditor ? GL.QUADS : GL.LINES);
-					ApplyWireMaterialDelegate.Invoke(CompareFunction.Always);
-					
-					const float quarter = 0.25f * blendBoxSize,
-						half = quarter * 2,
-						threeQuarters = half + quarter;
-					
-					Color lowGrey = LowGrey;
-					Color grey = Color.grey;
-					DrawLineFast(new Vector2(blendRect.x + quarter, blendRect.y), new Vector2(blendRect.x + quarter, blendRect.y + blendBoxSize), lowGrey);
-					DrawLineFast(new Vector2(blendRect.x + quarter, blendRect.y), new Vector2(blendRect.x + quarter, blendRect.y + blendBoxSize), lowGrey);
-					DrawLineFast(new Vector2(blendRect.x + threeQuarters, blendRect.y), new Vector2(blendRect.x + threeQuarters, blendRect.y + blendBoxSize), lowGrey);
-					DrawLineFast(new Vector2(blendRect.x, blendRect.y + quarter), new Vector2(blendRect.x + blendBoxSize, blendRect.y + quarter), lowGrey);
-					DrawLineFast(new Vector2(blendRect.x, blendRect.y + threeQuarters), new Vector2(blendRect.x + blendBoxSize, blendRect.y + threeQuarters), lowGrey);
-					DrawLineFast(new Vector2(blendRect.x + half, blendRect.y), new Vector2(blendRect.x + half, blendRect.y + blendBoxSize), grey);
-					DrawLineFast(new Vector2(blendRect.x, blendRect.y + half), new Vector2(blendRect.x + blendBoxSize, blendRect.y + half), grey);
-					GL.End();
-
-					GUI.Label(new Rect(blendRect.x + quarter, blendRect.y + half, blendBoxSize, 15), b2D.XLabel, EditorStyles.centeredGreyMiniLabel);
-					Matrix4x4 matrixP = GUI.matrix;
-					GUIUtility.RotateAroundPivot(-90, new Vector2(blendRect.x, blendRect.y) + new Vector2(half, half));
-					GUI.Label(new Rect(quarter, half - 15, blendBoxSize, 15), b2D.YLabel, EditorStyles.centeredGreyMiniLabel);
-					GUI.matrix = matrixP;
-
-					GL.Begin(Application.platform == RuntimePlatform.WindowsEditor ? GL.QUADS : GL.LINES);
-					ApplyWireMaterialDelegate.Invoke(CompareFunction.Always);
-					Vector2 circlePos = InverseLerp(b2D.Min, b2D.Max, property.vector2Value);
-					circlePos.y = 1 - circlePos.y;
-					circlePos *= blendBoxSize;
-					DrawCircleFast(blendRect.position + circlePos, CircleRadius, 2, CircleColor);
-					GL.End();
-				}
-
-#if UNITY_MATHEMATICS
-				bool IsFloat2() => property.propertyType == SerializedPropertyType.Generic && property.type == nameof(float2);
-
-				Vector2 v;
-				if (IsFloat2())
-				{
-					SerializedProperty x = property.FindPropertyRelative("x");
-					SerializedProperty y = property.FindPropertyRelative("y");
-					v = new Vector2(x.floatValue, y.floatValue);
-					// ReSharper disable once ConvertToUsingDeclaration
-					using (EditorGUI.ChangeCheckScope cC = new EditorGUI.ChangeCheckScope())
-					{
-						Rect labelRect = new Rect(blendBoxSize + 5, blendBoxSize / 2f - 50, Screen.width - blendBoxSize - 5, EditorGUIUtility.singleLineHeight);
-						EditorGUI.LabelField(labelRect, property.displayName, EditorStyles.boldLabel);
-						labelRect.y += 20;
-						EditorGUI.LabelField(labelRect, b2D.XLabel);
-						labelRect.y += EditorGUIUtility.singleLineHeight;
-						v.x = EditorGUI.DelayedFloatField(labelRect, v.x);
-						labelRect.y += EditorGUIUtility.singleLineHeight;
-						EditorGUI.LabelField(labelRect, b2D.YLabel);
-						labelRect.y += EditorGUIUtility.singleLineHeight;
-						v.y = EditorGUI.DelayedFloatField(labelRect, v.y);
-						if (cC.changed)
-						{
-							x.floatValue = Mathf.Clamp(v.x, b2D.Min.x, b2D.Max.x);
-							y.floatValue = Mathf.Clamp(v.y, b2D.Min.y, b2D.Max.y);
-						}
-					}
-
-					return;
-				}
-
-				v = property.vector2Value;
-#else
-				Vector2 v = property.vector2Value;
-#endif
-
-				using (EditorGUI.ChangeCheckScope cC = new EditorGUI.ChangeCheckScope())
-				{
-					Rect labelRect = new Rect(blendBoxSize + 5, blendBoxSize / 2f - 50, Screen.width - blendBoxSize - 5, EditorGUIUtility.singleLineHeight);
-					EditorGUI.LabelField(labelRect, property.displayName, EditorStyles.boldLabel);
-					labelRect.y += 20;
-					EditorGUI.LabelField(labelRect, b2D.XLabel);
-					labelRect.y += EditorGUIUtility.singleLineHeight;
-					v.x = EditorGUI.DelayedFloatField(labelRect, v.x);
-					labelRect.y += EditorGUIUtility.singleLineHeight;
-					EditorGUI.LabelField(labelRect, b2D.YLabel);
-					labelRect.y += EditorGUIUtility.singleLineHeight;
-					v.y = EditorGUI.DelayedFloatField(labelRect, v.y);
-					if (cC.changed)
-						property.vector2Value = new Vector2(Mathf.Clamp(v.x, b2D.Min.x, b2D.Max.x), Mathf.Clamp(v.y, b2D.Min.y, b2D.Max.y));
-				}
-			}
-			finally
-			{
-				GUI.EndGroup();
-				Profiler.EndSample();
-			}
-		}
 
 		public static void ShowContextMenu(SerializedProperty x, SerializedProperty y, Vector2 xLimit, Vector2 yLimit)
 		{
-			GenericMenu menu = new GenericMenu();
+			var menu = new GenericMenu();
 
 			menu.AddItem(new GUIContent("Center"), false, () =>
 			{
@@ -454,88 +272,5 @@ namespace Vertx.Attributes.Editor
 
 			menu.ShowAsContext();
 		}
-
-		static Vector2 InverseLerp(Vector2 min, Vector2 max, Vector2 value) =>
-			new Vector2(
-				Mathf.InverseLerp(min.x, max.x, value.x),
-				Mathf.InverseLerp(min.y, max.y, value.y));
-
-		static Vector2 Lerp(Vector2 min, Vector2 max, Vector2 value) =>
-			new Vector2(
-				Mathf.Lerp(min.x, max.x, value.x),
-				Mathf.Lerp(min.y, max.y, value.y));
-
-		private static void DrawLineFast(Vector2 from, Vector2 to, Color color)
-		{
-			// Align perfectly along a pixel
-			from.x = Mathf.Ceil(from.x) - 0.5f;
-			from.y = Mathf.Ceil(from.y) - 0.5f;
-			to.x = Mathf.Ceil(to.x) - 0.5f;
-			to.y = Mathf.Ceil(to.y) - 0.5f;
-			
-			GL.Color(color);
-			if (Application.platform == RuntimePlatform.WindowsEditor)
-			{
-				Vector2 tangent = (to - from).normalized;
-				Vector2 mult = new Vector2(tangent.y > tangent.x ? -1 : 1, tangent.y > tangent.x ? 1 : -1);
-				tangent = new Vector2(mult.x * tangent.y, mult.y * tangent.x) * 0.25f;
-				GL.Vertex(new Vector3(from.x + tangent.x, from.y + tangent.y, 0f));
-				GL.Vertex(new Vector3(from.x - tangent.x, from.y - tangent.y, 0f));
-				GL.Vertex(new Vector3(to.x - tangent.x, to.y - tangent.y, 0f));
-				GL.Vertex(new Vector3(to.x + tangent.x, to.y + tangent.y, 0f));
-			}
-			else
-			{
-				GL.Vertex(new Vector3(from.x, from.y, 0f));
-				GL.Vertex(new Vector3(to.x, to.y, 0f));
-			}
-		}
-
-		private const int circleDivisions = 18;
-
-		private static void DrawCircleFast(Vector2 position, float radius, float thickness, Color color)
-		{
-			GL.Color(color);
-			for (int i = 1; i <= circleDivisions; i++)
-			{
-				float vC = Mathf.PI * 2 * (i / (float)circleDivisions - 1 / (float)circleDivisions);
-				float vP = Mathf.PI * 2 * (i / (float)circleDivisions);
-				Vector2 from = position + new Vector2(Mathf.Sin(vP) * radius, Mathf.Cos(vP) * radius);
-				Vector2 to = position + new Vector2(Mathf.Sin(vC) * radius, Mathf.Cos(vC) * radius);
-				if (Application.platform == RuntimePlatform.WindowsEditor)
-				{
-					Vector2 tangent = (to - from).normalized;
-					Vector2 mult = new Vector2(tangent.y > tangent.x ? -1 : 1, tangent.y > tangent.x ? 1 : -1);
-					tangent = new Vector2(mult.x * tangent.y, mult.y * tangent.x) * (0.5f * thickness);
-					GL.Vertex(new Vector3(from.x + tangent.x, from.y + tangent.y, 0f));
-
-					GL.Vertex(new Vector3(to.x + tangent.x, to.y + tangent.y, 0f));
-					GL.Vertex(new Vector3(to.x - tangent.x, to.y - tangent.y, 0f));
-					GL.Vertex(new Vector3(from.x - tangent.x, from.y - tangent.y, 0f));
-				}
-				else
-				{
-					GL.Vertex(new Vector3(from.x, from.y, 0f));
-					GL.Vertex(new Vector3(to.x, to.y, 0f));
-				}
-			}
-		}
-
-
-		public override float GetPropertyHeight(SerializedProperty property, GUIContent label) => blendBoxSize;
-
-		private static Action<CompareFunction> s_ApplyWireMaterialDelegate;
-		private static Action<CompareFunction> ApplyWireMaterialDelegate
-			=> s_ApplyWireMaterialDelegate
-			   ?? (s_ApplyWireMaterialDelegate = (Action<CompareFunction>)Delegate.CreateDelegate(
-				   typeof(Action<CompareFunction>),
-				   typeof(HandleUtility).GetMethod(
-					   "ApplyWireMaterial",
-					   BindingFlags.NonPublic | BindingFlags.Static,
-					   null,
-					   new[] { typeof(CompareFunction) },
-					   null
-				   )
-			   ));
 	}
 }