+++ Editor/Controls/BetterEnumFlagsField.cs
@@ -12,23 +12,28 @@ namespace Vertx.Attributes.Editor
 	{
 		private delegate FieldInfo GetTypeFromPropertyBase(SerializedProperty property, out Type type);
 
-		private static GetTypeFromPropertyBase s_GetTypeFromProperty;
+		private static GetTypeFromPropertyBase? s_getTypeFromProperty;
 
 		private Type GetTypeFromProperty(SerializedProperty property)
 		{
-			if (s_GetTypeFromProperty == null)
+			if (s_getTypeFromProperty == null)
 			{
 				MethodInfo method = Type.GetType("UnityEditor.ScriptAttributeUtility,UnityEditor")!.GetMethod("GetFieldInfoFromProperty", BindingFlags.Static | BindingFlags.NonPublic)!;
-				s_GetTypeFromProperty = (GetTypeFromPropertyBase)Delegate.CreateDelegate(typeof(GetTypeFromPropertyBase), method);
+				s_getTypeFromProperty = (GetTypeFromPropertyBase)Delegate.CreateDelegate(typeof(GetTypeFromPropertyBase), method);
 			}
 
-			s_GetTypeFromProperty(property, out Type type);
+			s_getTypeFromProperty(property, out Type type);
 			return type;
 		}
 
 		public BetterEnumFlagsField(SerializedProperty property, FieldInfo fieldInfo, bool hideObsoleteNames)
 		{
-			EnumFlagsValueAndNames enumFlagsValueAndNames = EnumFlagsValueAndNames.Get(fieldInfo, hideObsoleteNames);
+			EnumFlagsValueAndNames? enumFlagsValueAndNames = EnumFlagsValueAndNames.Get(fieldInfo, hideObsoleteNames);
+			if (enumFlagsValueAndNames == null)
+			{
+				return;
+			}
+			
 			var dropdownButton = new DropdownButton(property.displayName, enumFlagsValueAndNames.GetName(property.intValue));
 			dropdownButton.RegisterClickCallback<(SerializedProperty property, EnumFlagsValueAndNames valuesAndNames)>(
 				(_, button, data) =>