+++ Runtime/ButtonAttribute.cs
@@ -6,7 +6,7 @@ namespace Vertx.Attributes
 	public sealed class ButtonAttribute : PropertyAttribute
 	{
 		public string MethodName { get; }
-		public Type StaticMethodType { get; }
+		public Type? StaticMethodType { get; }
 
 		public enum Location : byte
 		{
@@ -15,7 +15,7 @@ namespace Vertx.Attributes
 			Below
 		}
 		
-		public string DisplayNameOverride { get; set; } = null;
+		public string? DisplayNameOverride { get; set; }
 		public Location DisplayLocation { get; set; } = Location.Bottom;
 
 		public ButtonAttribute(string methodName)